refactor: move elevation gain calculation to static methods in IConverter
